pyproject.toml
[project]
name = "spotify-quiz-app"
version = "0.1.0"
description = "A quiz app for Spotify listening habits"
requires-python = ">=3.8"
dependencies = [
    "flask>=2.3.0",
    "spotipy>=2.23.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
